First working version of deploy machine code
The deploy machine is an HTTP server and a background daemon that
runs git-related actions whenever:

1. a post-receive hook is received on GitHub repos monitored by the
server
2. a periodic "check" script tells the daemon that the action should
be run.

In essence, for the Web site, the Web server reacts on "git push"
commands on the "master" branch to build a new version of the Web
site and on "git push" commands on the "build" branch to publish
the version to the production server.

Credentials should be set in environment variables. Typically, the
code more or less expects the environment to define a:

 JOSHFIREDEPLOY_KEY_DEPLOYMACHINE

... variable that contains the private SSH key of the deploy machine.
That key must have pull/push access to the repositories the deploy
machine interacts with.

Related to JOSHFIRE-22