One custom GIT_SSH script per private key, created on-the-fly
To use the private key set as environment variable in git commands,
the deploy machine needs to use a custom SSH script, exported in the
GIT_SSH environment variable. That script cannot take parameters.

In turn, the "ssh-noprompt.sh" script, used by default, relies on the
presence of the GIT_SSH_TMPKEY variable that sets the private SSH key
to use.

This worked fine for all first-level git commands. However, this
failed at the "npm install" when the repository that was cloned
depended on another private one using a "git+ssh" URL. It failed
because npm only exports the GIT_SSH variable and thus loses the
GIT_SSH_TMPKEY variable.

The deploy machine now generates a custom ssh script that hardcodes
the private key to use directly so as to only have to export the
GIT_SSH variable.