Added npm as explicit dependency
Heroku does not expose the "npm" utility once the code has been
deployed. However, the deploy machine needs to run "npm install" on
repositories it clones to be able to run the build/publication
scripts.

The deploy machine now has an explicit dependency on "npm" and
references "node_modules/.bin/npm" when it issues npm commands.