"repositories" folder reset on restart, version bumped to v0.9.3
As things stood, there was no simple way to reset the "repositories"
folder in environments such as Heroku without pushing a new version
of the server.

All that's needed from now on is a restart of the server. On Heroku,
this can be done with a fake config variable setting command such as:

 heroku config:set DEPLOYMACHINE_RESTART=1