First version of separation between Web server and scheduler
Beware, non working code! The goal is really just to park the code
in case we want to get back to it later on.

The approach will simply not work on Heroku: Web and scheduled dynos
have separate filesystems so cannot communicate using a simple file
queue. To implement things correctly, one would need to follow
Heroku background jobs practices:
https://devcenter.heroku.com/articles/background-jobs-queueing

In particular, there would need to be:
- one Web dyno
- one worker dyno
- a scheduler (the scheduler add-on should be enough)
- some message queue (e.g. the CloudAMQP add-on to pass tasks)

That seems overly complex for a simple deploy machine that will never
have to scale, so sticking to an all-in-one approach for the time
being.