Hook Machine! README completed, example added, MIT license
The "deploy machine" was renamed to "Hook Machine" with a nice
reference to Zouk machine's Maldon song (yup, I did say "nice").
The repository still needs to be renamed accordingly.

The README now details how to install and configure the Hook Machine.
It also features the concrete example of the Joshfire.com Web site.