stdout/stderr logged as they get updated during script execution
The server used to run Node's "exec" function that returns stdout
and stderr when the process is over. That was great and simple but
meant we did not know what was going on during the execution of the
script and had also no way to prevent memory usage from growing if
the script logged many things to the console.

The server now uses "spawn" and logs stdout and stderr on the fly,
on a line by line basis.